switch to signals for much state management
