+++ public/ts/list_elem_detailed.ts
@@ -1,12 +1,13 @@
 import { ListElementRenderer } from "./list_elem_renderer.js";
 import { fillAuthorDiv, getElementForCategory } from "./list_item_shared.js";
-import { CATEGORIES, fabric_category_id } from "./mod_search_logic.js";
+import { fabric_category_id } from "./initCategoriesSidebar.js";
 import { Mod } from "./mod_types.js";
 import { formatNumberCompact } from "./number_formatter.js";
 import {
     createCurseLinkIcon,
     createModrinthLinkIcon,
 } from "./platform_links.js";
+import { CATEGORIES } from "./initCategoriesSidebar.js"
 
 type ListElementTemplate = {
     li: HTMLLIElement;
@@ -135,7 +136,7 @@ class DetailedListElementRendererImpl extends ListElementRenderer<
                 // if not "Fabric"
                 if (category !== fabric_category_id) {
                     const catElem = document.createElement("li");
-                    catElem.textContent = CATEGORIES[category].name;
+                    catElem.textContent = CATEGORIES.get()[category].name;
                     categories.appendChild(catElem);
                 }
             }
